feat: implement find_local_files_by_extension function for enhanced file discovery
- Added the `find_local_files_by_extension` function to facilitate searching for local files by extension and optional name filtering, improving modularity and usability.
- Moved the function from `application_sdk.common.utils` to `application_sdk.activities.common.utils` to better organize utility functions.
- Updated relevant imports in the `Input` class and unit tests to reflect the new location of the function, ensuring consistency across the codebase.
