+++ application_sdk/inputs/__init__.py
@@ -2,9 +2,11 @@ import os
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, AsyncIterator, Iterator, List, Union
 
-from application_sdk.activities.common.utils import get_object_store_prefix
+from application_sdk.activities.common.utils import (
+    find_local_files_by_extension,
+    get_object_store_prefix,
+)
 from application_sdk.common.error_codes import IOError
-from application_sdk.common.utils import find_local_files_by_extension
 from application_sdk.constants import TEMPORARY_PATH
 from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.services.objectstore import ObjectStore
