+++ tests/unit/common/test_utils_file_discovery.py
@@ -3,7 +3,7 @@
 import tempfile
 from pathlib import Path
 
-from application_sdk.common.utils import find_local_files_by_extension
+from application_sdk.activities.common.utils import find_local_files_by_extension
 
 
 class TestFindFilesByExtension:
